Break circular references
