+++ python/python/Ice/EventLoopAdapter.py
@@ -1,9 +1,13 @@
 # Copyright (c) ZeroC, Inc.
 
+from __future__ import annotations
+
 from abc import ABC, abstractmethod
 from collections.abc import Awaitable, Coroutine
+from typing import TYPE_CHECKING
 
-from .Future import Future, FutureLike
+if TYPE_CHECKING:
+    import Ice
 
 
 class EventLoopAdapter(ABC):
@@ -12,7 +16,7 @@ class EventLoopAdapter(ABC):
     """
 
     @abstractmethod
-    def runCoroutine(self, coroutine: Coroutine) -> FutureLike:
+    def runCoroutine(self, coroutine: Coroutine) -> Ice.FutureLike:
         """
         Run a coroutine in the application configured event loop. The Ice run time will call this method to run
         coroutines returned by async dispatch methods. This method is called from the Ice dispatch thread.
@@ -30,7 +34,7 @@ class EventLoopAdapter(ABC):
         pass
 
     @abstractmethod
-    def wrapFuture(self, future: Future) -> Awaitable:
+    def wrapFuture(self, future: Ice.Future) -> Awaitable:
         """
         Wraps an Ice.Future so that it can be awaited in the application event loop. The Ice run time calls this method
         before returning a future to the application.
