+++ python/python/Ice/Util.py
@@ -40,7 +40,11 @@ def initialize(
 ) -> Ice.Communicator: ...
 
 
-def initialize(args=None, initData=None, configFile=None, eventLoop=None):
+def initialize(
+        args: list[str] | None = None,
+        initData: Ice.InitializationData | None = None,
+        configFile: str | None = None,
+        eventLoop: Ice.EventLoopAdapter | None = None) -> Ice.Communicator:
     """
     Creates a new communicator.
 
@@ -230,28 +234,6 @@ def getSliceDir() -> str | None:
             return dir
 
     return None
-
-
-_repr_running = threading.local()
-
-
-def safe_repr(obj: object) -> str:
-    if not hasattr(_repr_running, "set"):
-        _repr_running.set = set()
-
-    obj_id = id(obj)
-    if obj_id in _repr_running.set:
-        return "..."
-
-    _repr_running.set.add(obj_id)
-    try:
-        return repr(obj)
-    finally:
-        _repr_running.set.remove(obj_id)
-
-
-def format_fields(**fields: object) -> str:
-    return ", ".join(f"{k}={safe_repr(v)}" for k, v in fields.items())
 
 
 __all__ = ["initialize", "identityToString", "stringToIdentity", "createProperties", "getSliceDir"]
