+++ python/python/Ice/asyncio/EventLoopAdapter.py
@@ -1,10 +1,16 @@
 # Copyright (c) ZeroC, Inc.
 
+from __future__ import annotations
+
 import asyncio
 from collections.abc import Awaitable, Coroutine
+from typing import TYPE_CHECKING
 
 from ..EventLoopAdapter import EventLoopAdapter as Ice_EventLoopAdapter
-from ..Future import Future, FutureLike, wrap_future
+from ..Future import wrap_future
+
+if TYPE_CHECKING:
+    import Ice
 
 
 class EventLoopAdapter(Ice_EventLoopAdapter):
@@ -18,9 +24,9 @@ class EventLoopAdapter(Ice_EventLoopAdapter):
     def __init__(self, eventLoop: asyncio.AbstractEventLoop):
         self._eventLoop = eventLoop
 
-    def runCoroutine(self, coroutine: Coroutine) -> FutureLike:
+    def runCoroutine(self, coroutine: Coroutine) -> Ice.FutureLike:
         # Convert the concurrent.futures.Future to an awaitable by wrapping it with asyncio.wrap_future
         return asyncio.run_coroutine_threadsafe(coroutine, self._eventLoop)
 
-    def wrapFuture(self, future: Future) -> Awaitable:
+    def wrapFuture(self, future: Ice.Future) -> Awaitable:
         return wrap_future(future, loop=self._eventLoop)
